Initial commit of templated resume generator
Generating fake classes just so I can set __dict__ just so I can use
standard 'foo.bar' accessor syntax in jinja2 templates is probably a
terrible idea, but it makes templates _very_ easy to write!  While I'm
sure I'll run into problems with this (likely when I start trying nested
templates, e.g. a base.html or something), it's working okay for now.

This is super hacky right now though: nearly every path is hard-coded
And only plain-text is supported!  That will all need to change
eventually so I can actually use this.

There's a chance it'd be easier to blow this all away in favor of a
flask site, and using requests or something to call localhost?  Probably
not, because that's gross architecturally.  But possibly faster to get
running - flask has done most of this before!