Import sections into base layout instead of inheriting
Inheriting templates introduces a nightmare of block redefinition.
Since each section is really only defining one block, a bunch of extra
content (mostly headers and footers) gets duplicated.  Using a main
layout with {% import %} leads to a bit of recursion in these templates,
but none are self-recursive at least.

Something tells me this isn't the preferred way to do what I want with
jinja2, but it's very simple for now.