Implement education section for pretty-formatted .txt
I should eventually add the enrollment date, to be consistent with the
current convention of "dates on the left-hand-side of the timeline are
either start dates or date ranges".  That'll be a change in the .yaml
source and in each template.  Hooray, I'll get to see how maintenance of
these works!