Import barag_resume LaTeX class from old Resume repo
Now that resumes are generated on-the-fly, this class should probably be
in the generator package.

Unless I should use a git submodule for output/ ?  I'll work that out
later.  For now, committing the generated output seems fair to me.  It's
a bit unusual (you wouldn't commit compiled binaries), but also helps to
document the use of this package.

Old Resume repo: https://github.com/sjbarag/Resume